Done search vaccine by name

+++ src/main/java/com/swp_group03/vaccination/vaccination_schedule_children_tracking_project/model/response/Vaccine/ResponseVaccine.java
@@ -63,7 +63,8 @@ public class ResponseVaccine {
     LocalDate expirationDate;
 
 
-    BigDecimal price;
+    float price;
+
     String status;
 
 }
